+++ tests/testthat/test-download_pvolfiles.R
@@ -7,7 +7,6 @@ overwrite <- TRUE
 
 test_that("date input for download_pvolfiles() ", {
   skip_if_offline()
-  skip_if_no_aws.s3()
   # working with default
   expect_no_error(
     suppressMessages(
@@ -109,7 +108,6 @@ test_that("date input for download_pvolfiles() ", {
 
 test_that("Check radar code for download_pvolfiles() ", {
   skip_if_offline()
-  skip_if_no_aws.s3()
   expect_error(
     download_pvolfiles(date_min,
                        date_max,
@@ -134,7 +132,6 @@ test_that("Check radar code for download_pvolfiles() ", {
 
 test_that("Check path and overwrite for download_pvolfiles() ", {
   skip_if_offline()
-  skip_if_no_aws.s3()
   expect_error(
     download_pvolfiles(date_min, date_max, radars, 1, overwrite),
     "path is not a string (a length one character vector)",